requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
flask>=3.0.0
pandas>=2.0.0
